fix bug

+++ JSM/JSMServices/Services/OrderService.cs
@@ -198,8 +198,14 @@ public class OrderService : IOrderService
             var orderDetail = await
                 _orderDetailRepository.GetSingleWithAsync(c =>
                     c.OrderDetailId.ToUpper() == viewModel.OrderDetailID.ToUpper());
-            var warranty = await _warrantyRepository.GetSingleWithAsync(c => c.OrderDetailId == viewModel.OrderDetailID);
 
+            var warrantyList = _warrantyRepository.GetAll();
+            var warranty =
+                warrantyList.FirstOrDefault(c => c.OrderDetailId.ToUpper() == viewModel.OrderDetailID.ToUpper());
+            if (warranty == null)
+            {
+                warranty = new Warranty();
+            }
             var getOrder = await _orderRepository.GetAllWithAsync();
             order = getOrder.FirstOrDefault(c => c.OrderId.ToLower() == viewModel.OrderId.ToLower());
             if (order != null)
@@ -241,12 +247,20 @@ public class OrderService : IOrderService
 
             }
 
-            orderDetail.OrderDetailStatus = OrderDetail.OrderDetailStatuses.BuyBack;
-            await _orderDetailRepository.UpdateWithAsync(orderDetail);
-            _orderDetailRepository.SaveChanges();
+            if (orderDetail.OrderDetailStatus != OrderDetail.OrderDetailStatuses.BuyBack)
+            {
+                orderDetail.OrderDetailStatus = OrderDetail.OrderDetailStatuses.BuyBack;
+                await _orderDetailRepository.UpdateWithAsync(orderDetail);
+                _orderDetailRepository.SaveChanges();
+            }
+            
 
-            _warrantyRepository.Remove(warranty);
-            _warrantyRepository.SaveChanges();
+            if (warranty != null) 
+            {
+                _warrantyRepository.Remove(warranty);
+                _warrantyRepository.SaveChanges();
+            }
+            
 
             return order;
         }
